refactor(queue): unify typed producer and handler implementation with default JSON codec
- Remove typed_codec.go and typed_producer.go by integrating their functionality into producer.go and worker.go
- Add TaskProducer generic struct with methods to enqueue typed tasks using default JSON codec
- Add EnqueueFor and EnqueueForWithCodec generic functions for encoding payloads and enqueuing tasks
- Add HandleFor and HandleForWithCodec functions to decode task payloads and dispatch to typed handlers
- Add HandleTasker helper to register typed task handlers by task type
- Introduce defaultCodec using JSON codec implementation in codec.go
- Improve type safety and maintainability by consolidating task payload encoding/decoding logic
